+++ changelog.md
@@ -5,11 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [2.16.3] - 2026-06-29
+## [2.17.2] - 2026-06-29
 
 ### Fixed
 
 - `get_notes` now returns an empty tuple instead of raising `KeyError` for Nova cases that have no journal notes.
+
+## [2.17.1] - 2026-06-25
+
+### Fixed
+
+- Minor fixes in Boliglån modules.
+
+## [2.17.0] - 2026-06-10
+
+### Added
+
+- New modules to work with KMD Boliglån.
 
 ## [2.16.2] - 2026-03-13
 
@@ -282,6 +294,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release
 
+[2.17.1]: https://github.com/itk-dev-rpa/ITK-dev-shared-components/releases/tag/2.17.1
+[2.17.0]: https://github.com/itk-dev-rpa/ITK-dev-shared-components/releases/tag/2.17.0
+[2.16.2]: https://github.com/itk-dev-rpa/ITK-dev-shared-components/releases/tag/2.16.2
+[2.16.1]: https://github.com/itk-dev-rpa/ITK-dev-shared-components/releases/tag/2.16.1
 [2.16.0]: https://github.com/itk-dev-rpa/ITK-dev-shared-components/releases/tag/2.16.0
 [2.15.0]: https://github.com/itk-dev-rpa/ITK-dev-shared-components/releases/tag/2.15.0
 [2.14.0]: https://github.com/itk-dev-rpa/ITK-dev-shared-components/releases/tag/2.14.0